Fix bug in `SETTINGS` value serialization (#3)
### Summary

Clickhouse SQL has an optional [`SETTINGS` clause](settings-clause) with
`SELECT`s which can be used to configure [all kinds of options](settings).

In #1 and #2, I took a stab at implementing this, but the implementation contained a bug:
string setting values need to be single-quoted. Also, some settings require a `Map` and
some require a `Bool`. Let's support these too.

[settings-clause]: https://clickhouse.com/docs/en/sql-reference/statements/select#settings-in-select-query
[settings]: https://clickhouse.com/docs/en/operations/settings/settings

 ### Test Plan

Updated the unit test to cover the different `SETTINGS` data types.